Move ZBar.java to the com.sparrowwallet.sparrow.io package
